third_party/blink/web_tests/fast/dom/elementsFromPoint/elementsFromPoint-invalid-cases-expected.txt

PASS document.elementsFromPoint(300, 300)[0] is document.body.parentElement
PASS document.elementsFromPoint(199, 199)[0] is simpleDiv
PASS document.elementsFromPoint(199, 199)[1] is document.body
PASS document.elementsFromPoint(199, 199)[2] is document.body.parentElement
PASS document.elementsFromPoint(786, 1).length is 0
PASS document.elementsFromPoint(-1, 1).length is 0
PASS document.elementsFromPoint(-10, -10).length is 0
PASS successfullyParsed is true

TEST COMPLETE
